requirements.txt
pyserial
pyttsx3
libespeak-dev